refactor: filters
