+++ src/components/pages/events/events-board/filters/filters.jsx
@@ -1,13 +1,25 @@
 import PropTypes from 'prop-types';
-import React from 'react';
+import React, { useCallback } from 'react';
 
 import DropdownSelect from './dropdown-select';
 
-const Filters = ({ eventFilters, activeFilters, handleFilters }) => {
+const Filters = ({ eventFilters, activeFilters, setActiveFilters, setItemOffset }) => {
   const [type, region] = eventFilters;
   const { type: types, region: regions } = activeFilters;
   const isTypesSelected = types.length > 0;
   const isRegionsSelected = regions.length > 0;
+
+  const handleFilters = useCallback(
+    (filter, newValues) => {
+      const newFilters = {
+        ...activeFilters,
+        [filter.label]: newValues,
+      };
+      setActiveFilters(newFilters);
+      setItemOffset(0);
+    },
+    [activeFilters, setItemOffset, setActiveFilters]
+  );
 
   return (
     <div className="mt-8 flex gap-x-7 md:mt-14 [@media(max-width:550px)]:flex-col [@media(max-width:550px)]:space-y-5 [@media(max-width:550px)]:space-x-0">
@@ -38,7 +50,8 @@ Filters.propTypes = {
     type: PropTypes.arrayOf(PropTypes.string).isRequired,
     region: PropTypes.arrayOf(PropTypes.string).isRequired,
   }).isRequired,
-  handleFilters: PropTypes.func.isRequired,
+  setItemOffset: PropTypes.func.isRequired,
+  setActiveFilters: PropTypes.func.isRequired,
 };
 
 export default Filters;
